+++ cmd/tfproviderlint/tfproviderlint.go
@@ -22,7 +22,9 @@ import (
 	"github.com/bflad/tfproviderlint/passes/S003"
 	"github.com/bflad/tfproviderlint/passes/S004"
 	"github.com/bflad/tfproviderlint/passes/S005"
+	"github.com/bflad/tfproviderlint/passes/S006"
 	"github.com/bflad/tfproviderlint/passes/S007"
+	"github.com/bflad/tfproviderlint/passes/S008"
 )
 
 func main() {
@@ -40,6 +42,8 @@ func main() {
 		S003.Analyzer,
 		S004.Analyzer,
 		S005.Analyzer,
+		S006.Analyzer,
 		S007.Analyzer,
+		S008.Analyzer,
 	)
 }