+++ passes/S008/testdata/src/a/alias.go
@@ -1,0 +1,31 @@
+package a
+
+import (
+	s "github.com/hashicorp/terraform/helper/schema"
+)
+
+func falias() {
+	_ = s.Schema{ // want "schema of TypeList or TypeSet should not include Default"
+		Default: true,
+		Type:    s.TypeList,
+	}
+
+	_ = s.Schema{ // want "schema of TypeList or TypeSet should not include Default"
+		Default: true,
+		Type:    s.TypeSet,
+	}
+
+	_ = map[string]*s.Schema{
+		"name": { // want "schema of TypeList or TypeSet should not include Default"
+			Default: true,
+			Type:    s.TypeList,
+		},
+	}
+
+	_ = map[string]*s.Schema{
+		"name": { // want "schema of TypeList or TypeSet should not include Default"
+			Default: true,
+			Type:    s.TypeSet,
+		},
+	}
+}