+++ passes/S008/testdata/src/a/comment_ignore.go
@@ -1,0 +1,19 @@
+package a
+
+import (
+	"github.com/hashicorp/terraform/helper/schema"
+)
+
+func fcommentignore() {
+	//lintignore:S008
+	_ = schema.Schema{
+		Default: true,
+		Type:    schema.TypeList,
+	}
+
+	//lintignore:S008
+	_ = schema.Schema{
+		Default: true,
+		Type:    schema.TypeSet,
+	}
+}