+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 FEATURES
 
 * **New Check:** `S007`: check for `Schema` with both `Required` and `ConflictsWith` configured
+* **New Check:** `S008`: check for `Schema` of `TypeList` or `TypeSet` with `Default` configured
 
 BUG FIXES